+++ src/lib.rs
@@ -2,3 +2,6 @@ pub use wolf_engine_events as events;
 
 #[cfg(feature = "input")]
 pub use wolf_engine_input as input;
+
+#[cfg(feature = "window")]
+pub use wolf_engine_window as window;